Remove support for mono titles
With Roboto, I felt I needed the mono titles, especially in names with
"fi", because Roboto has a forced but inappropriate fi-ligature that is
ugly. But Inter is quite nice, and even function names in snake_case
look fine in it, so I will just keep it in the body font.