Disable contextual alternates to work around bug
I am not sure if it is a bug in Inter or in Chrome (and Firefox as
well). Both browsers affected surely looks like a problem with Inter,
but then again, regardless of what the intended state is, there is some
statefulnes in Chrome which is wrong. And in some places I can't
reproduce. So not sure what is going on, but disabling contextual
alternates fixes it, and I haven't noticed any things that it makes
worse, so let's disable calt.