Add support for section index pages
MkDocs supports index pages for a section. These are regular pages, and
they are part of the nav, but they have the "is_index" flag set. We use
the url of the first child of the section for the section link, and if
that child happens to be the index page, that makes the section link
point to the index page (this already worked like that). To prevent the
index from showing up in the aside nav, add a guard on "is_index" there.

Also exclude the page title from the breadcrumbs if the page is an index
... except for the root index, because the breadcrumbs look weird if
there is only a single breadcrumb equal to the page title.

Finally, the breadcrumb template was recursive to allow arbitrary
nesting, but the aside nav only allows sections and chapters (these
names are bad, I should just call them sections and pages). So that was
not necessary anyway, we can just inline it.