Fix nav spacing on narrow widths
I changed the section headers to have margins instead of paddings just
yet, in order for the right border to not stick out above for index
pages. But that had the unintend side effect of collapsing the margin
when the aside nav goes below the page. Add back that space.