Hide overflow in <aside> with ellipsis
In API docs, the titles can be long words that cannot be broken. In that
case, show an ellipsis, but don't overflow into the page.

Alternatively, we could turn on "overflow-wrap: break-word". That is
more informational, but also uglier. Ideally I would be able to break on
underscores, but that is not easy in this setup. The current affected
name has enough information in the left part that an ellipsis is okay.

Also fix <code> in <dd> appearance.