Use italic for <em>, not oblique
I noticed in Chromium, <em> is broken, and likely it has to do with this
slant. If I play around with the angle in the devtools then at some
point it does work. But I think the variable font is such that the
weight is the only axis, and slant is binary. So setting only the
font-style should be sufficient.